split up debug & physics context / providers, no more setup
