+++ packages/cannon-worker-api/package.json
@@ -28,11 +28,6 @@
     "prettier": "prettier --list-different .",
     "prettier-fix": "prettier --write ."
   },
-  "dependencies": {
-    "cannon-es": "^0.19.0",
-    "events": "^3.3.0",
-    "three": "^0.137.0"
-  },
   "devDependencies": {
     "@babel/core": "^7.16.5",
     "@babel/plugin-transform-runtime": "^7.16.5",
@@ -45,14 +40,17 @@
     "@types/three": "^0.137.0",
     "@typescript-eslint/eslint-plugin": "^5.7.0",
     "@typescript-eslint/parser": "^5.7.0",
+    "cannon-es": "^0.19.0",
     "eslint": "^8.5.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-es": "^4.1.0",
     "eslint-plugin-simple-import-sort": "^7.0.0",
     "eslint-plugin-typescript-enum": "^2.1.0",
+    "events": "^3.3.0",
     "prettier": "^2.5.1",
     "rollup": "^2.61.1",
     "rollup-plugin-web-worker-loader": "^1.6.1",
+    "three": "^0.137.0",
     "typescript": "^4.5.4"
   },
   "lint-staged": {
