+++ packages/react-three-cannon-examples/package.json
@@ -3,23 +3,6 @@
   "version": "1.0.1",
   "description": "Examples for @react-three/cannon",
   "private": true,
-  "keywords": [],
-  "dependencies": {
-    "@react-three/cannon": "*",
-    "@react-three/drei": "^8.11.1",
-    "@react-three/fiber": "^7.0.26",
-    "lerp": "^1.0.3",
-    "lodash-es": "^4.17.21",
-    "nice-color-palettes": "3.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-is": "^17.0.2",
-    "react-router-dom": "^6.2.1",
-    "styled-components": "^5.3.0",
-    "three": "^0.137.0",
-    "three-stdlib": "^2.8.8",
-    "zustand": "^3.6.7"
-  },
   "scripts": {
     "build": "tsc && vite build",
     "clean": "rm -rf dist/*",
@@ -36,7 +19,14 @@
     "not ie <= 11",
     "not op_mini all"
   ],
+  "dependencies": {
+    "@react-three/fiber": "^7.0.26",
+    "react": "^17.0.2",
+    "three": "^0.137.0"
+  },
   "devDependencies": {
+    "@react-three/cannon": "*",
+    "@react-three/drei": "^8.11.1",
     "@types/lodash-es": "^4.17.5",
     "@types/react": "^17.0.37",
     "@types/react-dom": "^17.0.11",
@@ -52,10 +42,19 @@
     "eslint-plugin-react": "^7.27.1",
     "eslint-plugin-simple-import-sort": "^7.0.0",
     "eslint-plugin-typescript-enum": "^2.1.0",
+    "lerp": "^1.0.3",
+    "lodash-es": "^4.17.21",
+    "nice-color-palettes": "3.0.0",
     "prettier": "^2.5.1",
+    "react-dom": "^17.0.2",
+    "react-is": "^17.0.2",
+    "react-router-dom": "^6.2.1",
+    "styled-components": "^5.3.0",
+    "three-stdlib": "^2.8.8",
     "typescript": "^4.5.4",
     "vite": "^2.7.3",
-    "vite-react-jsx": "^1.1.2"
+    "vite-react-jsx": "^1.1.2",
+    "zustand": "^3.6.7"
   },
   "lint-staged": {
     "*.{js,ts}": "eslint --cache --fix",
