+++ packages/react-three-cannon/package.json
@@ -30,8 +30,9 @@
     "prettier-fix": "prettier --write ."
   },
   "dependencies": {
-    "@pmndrs/cannon-worker-api": "^1.0.1",
-    "cannon-es-debugger": "^1.0.0"
+    "@react-three/fiber": "^7.0.26",
+    "react": ">=17.0",
+    "three": "^0.137.0"
   },
   "devDependencies": {
     "@babel/core": "^7.16.5",
@@ -39,7 +40,7 @@
     "@babel/preset-env": "^7.16.5",
     "@babel/preset-react": "^7.16.5",
     "@babel/preset-typescript": "^7.16.5",
-    "@react-three/fiber": "^7.0.26",
+    "@pmndrs/cannon-worker-api": "^1.0.1",
     "@rollup/plugin-babel": "^5.3.0",
     "@rollup/plugin-node-resolve": "^13.1.3",
     "@types/react": "^17.0.37",
@@ -47,6 +48,7 @@
     "@typescript-eslint/eslint-plugin": "^5.7.0",
     "@typescript-eslint/parser": "^5.7.0",
     "cannon-es": "^0.19.0",
+    "cannon-es-debugger": "^1.0.0",
     "eslint": "^8.5.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-es": "^4.1.0",
@@ -56,14 +58,10 @@
     "prettier": "^2.5.1",
     "react": "^17.0.2",
     "rollup": "^2.61.1",
-    "three": "^0.137.0",
     "typescript": "^4.5.4"
   },
   "lint-staged": {
     "*.{js,jsx,ts,tsx}": "eslint --cache --fix",
     "*.{js,jsx,ts,tsx,md}": "prettier --write"
-  },
-  "peerDependencies": {
-    "react": ">=17.0"
   }
 }
