+++ packages/react-three-cannon/src/hooks.ts
@@ -41,8 +41,9 @@ import type { DependencyList, MutableRefObject, Ref, RefObject } from 'react'
 import { useContext, useEffect, useLayoutEffect, useMemo, useRef, useState } from 'react'
 import { DynamicDrawUsage, Euler, InstancedMesh, MathUtils, Object3D, Quaternion, Vector3 } from 'three'
 
-import type { ProviderContext } from './setup'
-import { context, debugContext } from './setup'
+import { debugContext } from './debug-context'
+import type { PhysicsContext } from './physics-context'
+import { physicsContext } from './physics-context'
 
 export type AtomicApi<K extends AtomicName> = {
   set: (value: AtomicProps[K]) => void
@@ -111,7 +112,7 @@ let incrementingId = 0
 function subscribe<T extends SubscriptionName>(
   ref: RefObject<Object3D>,
   worker: CannonWorkerAPI,
-  subscriptions: ProviderContext['subscriptions'],
+  subscriptions: PhysicsContext['subscriptions'],
   type: T,
   index?: number,
   target: SubscriptionTarget = 'bodies',
@@ -136,7 +137,7 @@ function prepare(object: Object3D, props: BodyProps) {
 }
 
 function setupCollision(
-  events: ProviderContext['events'],
+  events: PhysicsContext['events'],
   { onCollide, onCollideBegin, onCollideEnd }: Partial<BodyProps>,
   uuid: string,
 ) {
@@ -158,7 +159,7 @@ function useBody<B extends BodyProps<unknown[]>>(
   deps: DependencyList = [],
 ): Api {
   const ref = useForwardedRef(fwdRef)
-  const { worker, refs, events, subscriptions } = useContext(context)
+  const { worker, refs, events, subscriptions } = useContext(physicsContext)
   const debugApi = useContext(debugContext)
 
   useLayoutEffect(() => {
@@ -187,7 +188,7 @@ function useBody<B extends BodyProps<unknown[]>>(
             object.setMatrixAt(i, temp.matrix)
             object.instanceMatrix.needsUpdate = true
             refs[id] = object
-            if (debugApi) debugApi.add(id, props, type)
+            debugApi?.add(id, props, type)
             setupCollision(events, props, id)
             return { ...props, args: argsFn(props.args) }
           })
@@ -195,7 +196,7 @@ function useBody<B extends BodyProps<unknown[]>>(
             const props = fn(i)
             prepare(object, props)
             refs[id] = object
-            if (debugApi) debugApi.add(id, props, type)
+            debugApi?.add(id, props, type)
             setupCollision(events, props, id)
             return { ...props, args: argsFn(props.args) }
           })
@@ -211,7 +212,7 @@ function useBody<B extends BodyProps<unknown[]>>(
     return () => {
       uuid.forEach((id) => {
         delete refs[id]
-        if (debugApi) debugApi.remove(id)
+        debugApi?.remove(id)
         delete events[id]
       })
       currentWorker.removeBodies({ uuid })
@@ -475,7 +476,7 @@ function useConstraint<T extends 'Hinge' | ConstraintTypes>(
   optns: ConstraintOptns | HingeConstraintOpts = {},
   deps: DependencyList = [],
 ): ConstraintORHingeApi<T> {
-  const { worker } = useContext(context)
+  const { worker } = useContext(physicsContext)
   const uuid = MathUtils.generateUUID()
 
   const refA = useForwardedRef(bodyA)
@@ -561,7 +562,7 @@ export function useSpring(
   optns: SpringOptns,
   deps: DependencyList = [],
 ): SpringApi {
-  const { worker } = useContext(context)
+  const { worker } = useContext(physicsContext)
   const [uuid] = useState(() => MathUtils.generateUUID())
 
   const refA = useForwardedRef(bodyA)
@@ -597,7 +598,7 @@ function useRay(
   callback: (e: RayhitEvent) => void,
   deps: DependencyList = [],
 ) {
-  const { worker, events } = useContext(context)
+  const { worker, events } = useContext(physicsContext)
   const [uuid] = useState(() => MathUtils.generateUUID())
   useEffect(() => {
     events[uuid] = { rayhit: callback }
@@ -661,7 +662,7 @@ export function useRaycastVehicle(
   deps: DependencyList = [],
 ): [RefObject<Object3D>, RaycastVehiclePublicApi] {
   const ref = useForwardedRef(fwdRef)
-  const { worker, subscriptions } = useContext(context)
+  const { worker, subscriptions } = useContext(physicsContext)
 
   useLayoutEffect(() => {
     if (!ref.current) {
@@ -731,7 +732,7 @@ export function useContactMaterial(
   options: ContactMaterialOptions,
   deps: DependencyList = [],
 ): void {
-  const { worker } = useContext(context)
+  const { worker } = useContext(physicsContext)
   const [uuid] = useState(() => MathUtils.generateUUID())
 
   useEffect(() => {
