+++ packages/react-three-cannon/src/physics-context.ts
@@ -1,0 +1,28 @@
+import type {
+  CannonWorkerAPI,
+  CollideBeginEvent,
+  CollideEndEvent,
+  CollideEvent,
+  RayhitEvent,
+  Refs,
+  Subscriptions,
+} from '@pmndrs/cannon-worker-api'
+import type { MutableRefObject } from 'react'
+import { createContext } from 'react'
+
+type CannonEvent = CollideBeginEvent | CollideEndEvent | CollideEvent | RayhitEvent
+type CallbackByType<T extends { type: string }> = {
+  [K in T['type']]?: T extends { type: K } ? (e: T) => void : never
+}
+
+export type CannonEvents = { [uuid: string]: Partial<CallbackByType<CannonEvent>> }
+
+export type PhysicsContext = {
+  bodies: MutableRefObject<{ [uuid: string]: number }>
+  events: CannonEvents
+  refs: Refs
+  subscriptions: Subscriptions
+  worker: CannonWorkerAPI
+}
+
+export const physicsContext = createContext<PhysicsContext>({} as PhysicsContext)
